+++ Source/Core/Core/Core.vcxproj
@@ -59,7 +59,6 @@
     <ClCompile Include="DSP\DSPCaptureLogger.cpp" />
     <ClCompile Include="DSP\DSPCodeUtil.cpp" />
     <ClCompile Include="DSP\DSPCore.cpp" />
-    <ClCompile Include="DSP\DSPEmitter.cpp" />
     <ClCompile Include="DSP\DSPHWInterface.cpp" />
     <ClCompile Include="DSP\DSPMemoryMap.cpp" />
     <ClCompile Include="DSP\DSPStacks.cpp" />
@@ -72,6 +71,7 @@
     <ClCompile Include="DSP\Interpreter\DSPIntLoadStore.cpp" />
     <ClCompile Include="DSP\Interpreter\DSPIntMisc.cpp" />
     <ClCompile Include="DSP\Interpreter\DSPIntMultiplier.cpp" />
+    <ClCompile Include="DSP\Jit\DSPEmitter.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitArithmetic.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitBranch.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitCCUtil.cpp" />
@@ -286,7 +286,6 @@
     <ClInclude Include="DSP\DSPCodeUtil.h" />
     <ClInclude Include="DSP\DSPCommon.h" />
     <ClInclude Include="DSP\DSPCore.h" />
-    <ClInclude Include="DSP\DSPEmitter.h" />
     <ClInclude Include="DSP\DSPHost.h" />
     <ClInclude Include="DSP\DSPHWInterface.h" />
     <ClInclude Include="DSP\DSPMemoryMap.h" />
@@ -296,6 +295,7 @@
     <ClInclude Include="DSP\Interpreter\DSPInterpreter.h" />
     <ClInclude Include="DSP\Interpreter\DSPIntExtOps.h" />
     <ClInclude Include="DSP\Interpreter\DSPIntUtil.h" />
+    <ClInclude Include="DSP\Jit\DSPEmitter.h" />
     <ClInclude Include="DSP\Jit\DSPJitRegCache.h" />
     <ClInclude Include="DSP\LabelMap.h" />
     <ClInclude Include="ec_wii.h" />
