+++ Source/Core/Core/DSP/DSPCore.h
@@ -13,7 +13,8 @@
 
 #include "Core/DSP/DSPBreakpoints.h"
 #include "Core/DSP/DSPCaptureLogger.h"
-#include "Core/DSP/DSPEmitter.h"
+
+class DSPEmitter;
 
 enum : u32
 {
