+++ Source/Core/Core/DSP/DSPTables.cpp
@@ -8,9 +8,9 @@
 
 #include "Common/CommonTypes.h"
 
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/Interpreter/DSPIntExtOps.h"
 #include "Core/DSP/Interpreter/DSPInterpreter.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 // clang-format off
 const DSPOPCTemplate opcodes[] =
