+++ Source/Core/Core/DSP/Jit/DSPJitArithmetic.cpp
@@ -6,8 +6,8 @@
 
 #include "Common/CommonTypes.h"
 
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPMemoryMap.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
