+++ Source/Core/Core/DSP/Jit/DSPJitCCUtil.cpp
@@ -5,7 +5,7 @@
 // Additional copyrights go to Duddie and Tratax (c) 2004
 
 #include "Core/DSP/DSPCore.h"
-#include "Core/DSP/DSPEmitter.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
