+++ Source/Core/Core/DSP/Jit/DSPJitExtOps.cpp
@@ -4,8 +4,8 @@
 
 #include "Common/CommonTypes.h"
 
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPMemoryMap.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
