+++ Source/Core/Core/DSP/Jit/DSPJitLoadStore.cpp
@@ -6,9 +6,9 @@
 
 #include "Common/CommonTypes.h"
 
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPMemoryMap.h"
 #include "Core/DSP/Interpreter/DSPInterpreter.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
