+++ Source/Core/Core/DSP/Jit/DSPJitMultiplier.cpp
@@ -9,7 +9,7 @@
 #include "Common/CommonTypes.h"
 
 #include "Core/DSP/DSPCore.h"
-#include "Core/DSP/DSPEmitter.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
