+++ Source/Core/Core/DSP/Jit/DSPJitRegCache.cpp
@@ -9,8 +9,8 @@
 #include "Common/Assert.h"
 #include "Common/Logging/Log.h"
 
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPMemoryMap.h"
+#include "Core/DSP/Jit/DSPEmitter.h"
 
 using namespace Gen;
